requirements.txt
pytest==6.2.4
#pytest-dependency==0.5.1
#pytest-forked==1.3.0
pytest-html==3.1.1
#pytest-metadata==1.11.0
#pytest-mock==3.6.1
#pytest-rerunfailures==10.0
pytest-selenium==3.0.0
pytest-sugar==0.9.5
pytest-variables==1.9.0
#pytest-xdist==2.3.0
#pytest-cov==2.12.1
#coverage==5.5

